requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
openai==1.51.2
httpx<0.28.0
pytesseract==0.3.13
Pillow==10.4.0
scikit-learn==1.5.2
numpy==2.1.1
